+++ plugins/wellness/__init__.py
@@ -1,0 +1,10 @@
+from plugins.base import Plugin
+
+
+class WellnessPlugin(Plugin):
+    id = "wellness"
+    name = "Wellness"
+    version = "1.0.0"
+    author = "Helix"
+    description = "Pengingat kesehatan (minum air, istirahat, dll)"
+    module_name = "wellness"